Tests/LibWeb/Text/expected/HTML/Window-onerror.txt

onerror event fired: Error: Uncaught error
